backend/auth-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
slowapi==0.1.9
redis==5.0.1
